modify:data_type

+++ 01_system/mainInit_2803x.c
@@ -360,7 +360,7 @@ void InitPieCtrl(void)
 /************************************************************
 所有误操作的中断处理
 ************************************************************/
-int gErrorIntCnt;
+s16 gErrorIntCnt;
 #ifdef TARGET_GS32
 __interrupt void rsvd_ISR(void)
 #else
@@ -469,7 +469,7 @@ void InitSetGpio(void)
 ************************************************************/
 void InitSetAdc(void)
 {
-    int waite;
+    s16 waite;
     
     EALLOW;
     AdcRegs.ADCCTL1.bit.ADCBGPWD  = 1;      // Power ADC BG
